Move konoha.konoha_token.Token to konoha.data.token.Token
