+++ konoha/word_tokenizers/kytea_tokenizer.py
@@ -1,6 +1,6 @@
 from typing import Optional
 
-from konoha.konoha_token import Token
+from konoha.data.token import Token
 from konoha.word_tokenizers.tokenizer import BaseTokenizer
