+++ konoha/word_tokenizers/mecab_tokenizer.py
@@ -1,7 +1,7 @@
 from typing import Optional
 from typing import List
 
-from konoha.konoha_token import Token
+from konoha.data.token import Token
 from konoha.word_tokenizers.tokenizer import BaseTokenizer
