+++ konoha/word_tokenizers/sudachi_tokenizer.py
@@ -1,4 +1,4 @@
-from konoha.konoha_token import Token
+from konoha.data.token import Token
 from konoha.word_tokenizers.tokenizer import BaseTokenizer
